# openapi/argus.v1.yaml
openapi: 3.1.0
info:
  title: Argus Phase-1 API
  version: 0.1.0
  description: >
    Minimal API surface for the Phase-1 walking skeleton: local auth, enrollment tokens,
    collector registry, and one metric query. Conventions (problem+json, cursor pagination,
    X-Request-ID, CSRF double-submit, Idempotency-Key) follow the architecture package §22.
    Session auth uses the `argus_session` cookie; unsafe methods additionally require
    `X-CSRF-Token` matching the `argus_csrf` cookie.
    Inventory operations additionally declare `x-argus-capability` (required RBAC-SC
    capability) and `x-argus-scope` (org|site|device_group|device); the contract test
    compares these against the route registry so an implemented inventory route without
    enforcement metadata (or with mismatched metadata) fails CI. Credential operations
    declare the same extensions (scope `org`; the binding target scope is validated
    server-side). Credentials are write-only: no route, response, or query returns
    secret material — `GET` returns metadata only.
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Local login (org slug + email + password)
      security: []          # public; rate-limited 10/min/IP
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Session established; Set-Cookie argus_session + argus_csrf
          headers:
            Set-Cookie: { schema: { type: string }, description: "argus_session=...; HttpOnly; SameSite=Lax; Path=/ ; argus_csrf=... (readable)" }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Revoke current session
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: Revoked }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/me:
    get:
      tags: [identity]
      summary: Current user, organization, role
      parameters: [{ $ref: '#/components/parameters/RequestId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /v1/sites:
    get:
      tags: [tenancy]
      summary: List sites of the current organization
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Site' } }
                  next_cursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /v1/enrollments:
    get:
      tags: [collectors]
      summary: List enrollment token metadata (never raw tokens)
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/EnrollmentToken' } }
                  next_cursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [collectors]
      summary: Create a one-time enrollment token (raw token returned exactly once)
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollmentTokenCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrollmentTokenCreated' }
        '400': { $ref: '#/components/responses/Validation' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/collectors:
    get:
      tags: [collectors]
      summary: List collectors (org-scoped)
      parameters:
        - { name: 'filter[site_id]', in: query, schema: { type: string, format: uuid } }
        - { name: 'filter[status]', in: query, schema: { type: string, enum: [pending, active, stale, revoked] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Collector' } }
                  next_cursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /v1/collectors/{id}:
    get:
      tags: [collectors]
      summary: Collector detail (identity, status, policy, certificate, spool stats)
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CollectorDetail' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/collectors/{id}/revoke:
    post:
      tags: [collectors]
      summary: Revoke a collector (idempotent)
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '200':
          description: Revoked (or already revoked)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Collector' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/collectors/{id}/policy:resync:
    post:
      tags: [collectors]
      summary: Issue the next policy version for a collector
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '202':
          description: Accepted; policy version incremented and queued for delivery
          content:
            application/json:
              schema:
                type: object
                required: [policy_version]
                properties: { policy_version: { type: integer } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/devices:
    get:
      tags: [inventory]
      summary: List devices (cursor pagination, soft-deleted hidden by default)
      x-argus-capability: device.read
      x-argus-scope: site
      parameters:
        - { name: 'filter[site_id]', in: query, schema: { type: string, format: uuid } }
        - { name: 'filter[status]', in: query, schema: { type: string, enum: [new, up, down, degraded, maintenance, retired] } }
        - { name: 'filter[kind]', in: query, schema: { type: string } }
        - { name: include_deleted, in: query, schema: { type: boolean, default: false } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Device' } }
                  next_cursor: { type: [string, 'null'] }
                  has_more: { type: boolean }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [inventory]
      summary: Add a manual device (admin)
      x-argus-capability: device.write
      x-argus-scope: site
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/devices/{id}:
    get:
      tags: [inventory]
      summary: Device detail
      x-argus-capability: device.read
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: include_deleted, in: query, schema: { type: boolean, default: false } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [inventory]
      summary: Update a device (admin; omitted fields unchanged, null clears nullable fields)
      x-argus-capability: device.write
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [inventory]
      summary: Soft-delete a device (admin)
      x-argus-capability: device.write
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Soft-deleted }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/devices/{id}/identity-history:
    get:
      tags: [inventory]
      summary: Identity attribute history of a device (open rows have null last_seen_at)
      x-argus-capability: device.identity.read
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/IdentityRecord' } }
                  next_cursor: { type: [string, 'null'] }
                  has_more: { type: boolean }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/devices/{id}/merge:
    post:
      tags: [inventory]
      summary: 'Merge source devices into {id}: re-points identity history + interfaces, soft-deletes sources (admin)'
      x-argus-capability: device.merge
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceMergeRequest' }
      responses:
        '200':
          description: Merged; target device returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/devices/{id}/split:
    post:
      tags: [inventory]
      summary: 'Split identity-history rows off {id} into a new device (admin)'
      x-argus-capability: device.split
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceSplitRequest' }
      responses:
        '201':
          description: New device created; identity rows re-pointed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/devices/{id}/interfaces:
    get:
      tags: [inventory]
      summary: List interfaces of a device
      x-argus-capability: interface.read
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Interface' } }
                  next_cursor: { type: [string, 'null'] }
                  has_more: { type: boolean }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [inventory]
      summary: Add a manual interface to a device (admin; UQ(device_id, if_index))
      x-argus-capability: interface.write
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InterfaceCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interface' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/interfaces/{id}:
    get:
      tags: [inventory]
      summary: Interface detail
      x-argus-capability: interface.read
      x-argus-scope: device
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interface' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [inventory]
      summary: Update editable interface metadata (admin; if_index is immutable)
      x-argus-capability: interface.write
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InterfaceUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interface' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [inventory]
      summary: Delete an interface (admin; the 000008 shape has no deleted_at)
      x-argus-capability: interface.write
      x-argus-scope: device
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/device-groups:
    get:
      tags: [inventory]
      summary: List device groups
      x-argus-capability: device_group.read
      x-argus-scope: device_group
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/DeviceGroup' } }
                  next_cursor: { type: [string, 'null'] }
                  has_more: { type: boolean }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [inventory]
      summary: Create a dynamic device group (admin; selector is a jsonb object)
      x-argus-capability: device_group.write
      x-argus-scope: org
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceGroupCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceGroup' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /v1/credentials/{id}:
    get:
      tags: [credentials]
      summary: Credential metadata detail (metadata only; there is no secret read)
      x-argus-capability: credential.read_metadata
      x-argus-scope: org
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: OK (metadata only)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CredentialMetadata' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: argus_session
...
  responses:
    Validation:
      description: Validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthenticated:
      description: Not authenticated
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Insufficient role or CSRF failure
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    LoginRequest:
      type: object
      required: [org_slug, email, password]
      properties:
        org_slug: { type: string, pattern: '^[a-z0-9][a-z0-9-]{1,62}$' }
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    Me:
      type: object
      required: [user, org]
      properties:
        user: { $ref: '#/components/schemas/User' }
        org: { $ref: '#/components/schemas/Org' }
...
security:
  - sessionCookie: []